% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterByPlotDF.R
\name{filterByPlotDF}
\alias{filterByPlotDF}
\title{Funkcja znajdująca jednotematyczne wątki}
\usage{
filterByPlotDF(dateStart, dateStop, dane, npost = 5, freq = 0.5)
}
\arguments{
\item{dateStart}{data początku okresu w formacie "yyyy-mm-dd" jako character}

\item{dateStop}{data końca okresu w formacie "yyyy-mm-dd" jako character}

\item{dane}{ramka danych z kolumnami: created_at, id, thread_id, rzeczownik}

\item{npost}{minimalna liczba postów w wątku które są uwzględniane}

\item{freq}{minimalna częstotliwośc wytstępowania słowa na post}
}
\value{
zwraca ramek danych wszystkich wybranych słów z ich częstościami
}
\description{
Funkcja znajduje wątki w których częściej niż f występują słowa na post.
}

