% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterByPlotList.R
\name{filterByPlotList}
\alias{filterByPlotList}
\title{Funkcja znajdująca jednotematyczne wątki}
\usage{
filterByPlotList(dateStart, dateStop, dane, npost = 5, freq = 0.5)
}
\arguments{
\item{dateStart}{data początku okresu w formacie "yyyy-mm-dd" jako character}

\item{dateStop}{data końca okresu w formacie "yyyy-mm-dd" jako character}

\item{dane}{ramka danych z kolumnami: created_at, id, thread_id, rzeczownik}

\item{npost}{minimalna liczba postów w wątku które są uwzględniane}

\item{freq}{minimalna częstotliwośc wytstępowania słowa na post}
}
\value{
zwraca listę ramek danych, każdy element listy jest nazwany numerem wątku (thread_id)
}
\description{
Funkcja znajduje wątki w których częściej niż f występują jakieś rzeczowniki per post. Dla każdego takiego wątku zwracana oddzielna ramka danych w liście o nazwie tread_id wątku.
}

