% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalDFrzecz.R
\name{finalDFRzecz}
\alias{finalDFRzecz}
\title{Funkcja oddzielająca słowa}
\usage{
finalDFRzecz(dane)
}
\arguments{
\item{dane}{- ramka danych do przekształcenia, musi zwierać kolumny o nazwie: id, parent_id, tread_id, created_at, user_name, body, rzeczowniki.}
}
\value{
ramkę danych z kolumnami: id, parent_id, tread_id, created_at, user_name, body, rzeczownik przy czym w kolumnie rzeczownik występująpojedyńcze rzeczowniki bez powtórzeń w ramch jednego postu.
}
\description{
Funkcja przekształca wejściową ramkę danych na dane z oddzielonymi poszczegulnymi słowami z kolumny rzeczowniki w kolumnie o nazwie rzeczownik. Ponadto zachowuje tylko kolumny o nazwach: id, parent_id, tread_id, created_at, user_name, body. Ponadto funkcja usówa stopwordsy i zamienia czasowniki na rzeczowniki odczasownikowe.
}

