% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_pack.R
\name{fun_pack}
\alias{fun_pack}
\title{Obliczenia emisji spalin}
\usage{
fun_pack(
  dane = input,
  kategoria = "Passenger Cars",
  euro = "Euro 5",
  mode = "",
  substancja = c("CO", "EC")
)
}
\arguments{
\item{substancja}{}
}
\value{

}
\description{
Obliczenia emisji spalin
}
