% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Greeting function}
\usage{
hello(x)
}
\arguments{
\item{x}{the name of the user.}
}
\description{
The greeting function greets the user. Gagat
}
\details{
This function uses \code{\link[base]{paste}}.
}
