% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_mean.R
\name{list_mean}
\alias{list_mean}
\title{Mean of list}
\usage{
list_mean(x)
}
\arguments{
\item{x}{list of vectors}
}
\description{
Computes bardzo ładnie mean of vectors
}
\details{
This function uses \code{\link[base]{lapply}} and  \code{\link[base]{mean}} .
}
