% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listMean.R
\name{listMean}
\alias{listMean}
\title{Mean of list elements}
\usage{
listMean(x)
}
\arguments{
\item{x}{list of vectors}
}
\description{
Calculates mean of each vector in a list
}
\details{
This function uses \code{\link[base]{mean}}.
}
