% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poprawaOrtografii.R
\name{poprawaOrtografii}
\alias{poprawaOrtografii}
\title{Poprawianiwe Ortografii}
\usage{
poprawaOrtografii(body)
}
\arguments{
\item{body}{napis, który ma zostać poprawiony}
}
\value{
zwraca poprawiony napis (wektor 1-elem. typu character)
}
\description{
Funkcja służy do poprawy ortografii w podanych tekstach
}

