% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separateNouns.R
\name{separateNouns}
\alias{separateNouns}
\title{Nowa ramka danych}
\usage{
separateNouns(dane)
}
\arguments{
\item{dane}{- ramka danych o kolumnach X, id, parent_id, created_at, thread_id, subject, user_name, user_link, source, body, rzeczowniki, gdzie rzeczowniki są podane po kolei oddzielone |}
}
\value{
zwraca podobną danych w której zamiast rzeczowników jest rzeczownik - pojedyńcze słowo oraz nie ma kolumny X
}
\description{
Funkcja tworząca ramke danych z pojedyńczymi rzeczownikami
}

