% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usuwanieStopwordsow.R
\name{usuwanieStopwordsow}
\alias{usuwanieStopwordsow}
\title{Usuwanie stopwordsów}
\usage{
usuwanieStopwordsow(wektor)
}
\arguments{
\item{wektor}{wektor napisow postaci wyraz1|wyraz2|wyraz3}
}
\value{
zwraca zmieniony wektor w tej samej formie (zwroty wielowyrazowe 
laczy znakiem _ zamiast spacji) ale bez stopwordsow
}
\description{
Funkcja służy do usuwania tzw. stopwordsow
}

