% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wydobywanieRzeczownikow.R
\name{wydobywanieRzeczownikow}
\alias{wydobywanieRzeczownikow}
\title{Wydobywanie Rzeczowników}
\usage{
wydobywanieRzeczownikow(body)
}
\arguments{
\item{body}{napis, który ma zostać poprawiony}
}
\value{
zwraca poprawiony napis (wektor 1-elem. typu character)
}
\description{
Funkcja służy do wybierania rzeczowników z podanego napisu
}

