% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zamianaCzasownikow.R
\name{zamianaCzasownikow}
\alias{zamianaCzasownikow}
\title{Zamiana Czasowników}
\usage{
zamianaCzasownikow(wektor)
}
\arguments{
\item{wektor}{wektor napisow postaci wyraz1|wyraz2|wyraz3}
}
\value{
zwraca zmieniony wektor w tej samej formie (zwroty wielowyrazowe laczy znakiem _ zamiast spacji)
}
\description{
Funkcja służy do zamiany czasowników w bezokoliczniku na odpowiadające im
rzeczowniki odczasownikowe (czasowniki typu: pływać -> pływanie)
}

